backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
python-json-logger>=2.0.0
prometheus-client>=0.19.0
redis>=5.0.0
cryptography>=40.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
psutil>=5.9.0
snowflake-connector-python>=3.0.0
oracledb>=2.0.0
pyodbc>=5.0.0
python-socketio>=5.10.0
kafka-python>=2.3.0